AI-Agent-Automation/requirements.txt
selenium==4.16.0
webdriver-manager==4.0.2
pyttsx3==2.99
TestLink-API-Python-client==0.8.1
pyautogui==0.9.54
openai
python-dotenv



